refactor(core): adopt shared error-response schema factory
Move errorResponseSchema/errorResponseZodSchema out of base.ts into a
dedicated endpoints/responses.ts and add an errorResponses() multi-status
helper. Replace 26 inlined `{ success: false, error: {...} }` response
blocks across 16 endpoint files with errorResponseSchema(description).

The custom-detail error block in delete.ts (409 with { relation, count })
is intentionally left inline. 917 tests pass; typecheck + biome clean.
