+++ packages/core/src/endpoints/aggregate.ts
@@ -12,6 +12,7 @@ import type {
 } from '../core/types';
 import { parseAggregateQuery } from '../core/types';
 import { CrudEndpoint } from './base';
+import { errorResponseSchema } from './responses';
 
 /**
  * Default aggregate configuration.
@@ -137,21 +138,7 @@ export abstract class AggregateEndpoint<
             },
           },
         },
-        400: {
-          description: 'Invalid aggregation request',
-          content: {
-            'application/json': {
-              schema: z.object({
-                success: z.literal(false),
-                error: z.object({
-                  code: z.string(),
-                  message: z.string(),
-                  details: z.unknown().optional(),
-                }),
-              }),
-            },
-          },
-        },
+        400: errorResponseSchema('Invalid aggregation request'),
       },
     };
   }
