+++ packages/core/src/endpoints/base.ts
@@ -704,51 +704,3 @@ export abstract class CrudEndpoint<
     return data;
   }
 }
-
-// ============================================================================
-// Shared error-response Zod schema factory
-// ============================================================================
-
-/**
- * Build a `{ success: false, error: { code, message, details? } }` Zod
- * response schema. Replaces the 14+ inlined copies of this schema across
- * endpoint files.
- */
-export function errorResponseSchema(description?: string) {
-  const schema = z.object({
-    success: z.literal(false),
-    error: z.object({
-      code: z.string(),
-      message: z.string(),
-      details: z.unknown().optional(),
-    }),
-  });
-  return {
-    description: description ?? 'Error',
-    content: {
-      'application/json': { schema },
-    },
-  };
-}
-
-/**
- * Returns just the Zod schema (for callers that want to embed it in a
- * custom response shape).
- */
-export function errorResponseZodSchema(): ZodObject<{
-  success: z.ZodLiteral<false>;
-  error: ZodObject<{
-    code: z.ZodString;
-    message: z.ZodString;
-    details: z.ZodOptional<z.ZodUnknown>;
-  }>;
-}> {
-  return z.object({
-    success: z.literal(false),
-    error: z.object({
-      code: z.string(),
-      message: z.string(),
-      details: z.unknown().optional(),
-    }),
-  });
-}
