+++ packages/core/src/endpoints/batch-upsert.ts
@@ -5,6 +5,7 @@ import { getManagedInputExclusions, rethrowAsConstraintError } from '../core/man
 import type { HookMode, MetaInput, OpenAPIRouteSchema } from '../core/types';
 import { applyComputedFields } from '../core/types';
 import { CrudEndpoint } from './base';
+import { errorResponseSchema } from './responses';
 import { type ModelObject, getSchemaFields } from './types';
 
 /**
@@ -247,21 +248,7 @@ export abstract class BatchUpsertEndpoint<
             },
           },
         },
-        400: {
-          description: 'Validation error',
-          content: {
-            'application/json': {
-              schema: z.object({
-                success: z.literal(false),
-                error: z.object({
-                  code: z.string(),
-                  message: z.string(),
-                  details: z.unknown().optional(),
-                }),
-              }),
-            },
-          },
-        },
+        400: errorResponseSchema('Validation error'),
       },
     };
   }
